refactor(guardrails): extract shared tool wrap logic into base mixin
Move _extract_tool_input_data and the entire PRE+POST wrap_tool_call
implementation into BuiltInGuardrailMiddlewareMixin._create_tool_wrap_hook,
eliminating duplication between UiPathPIIDetectionMiddleware and
UiPathHarmfulContentMiddleware.
