+++ src/uipath_langchain/guardrails/middlewares/_base.py
@@ -5,7 +5,9 @@ import json
 import logging
 from typing import Any
 
+from langchain.agents.middleware import AgentMiddleware, wrap_tool_call
 from langchain_core.messages import AIMessage, BaseMessage, HumanMessage, ToolMessage
+from langgraph.prebuilt.tool_node import ToolCallRequest
 from langgraph.types import Command
 from uipath.core.guardrails import (
     GuardrailValidationResult,
@@ -13,12 +15,19 @@ from uipath.core.guardrails import (
 )
 from uipath.platform import UiPath
 from uipath.platform.guardrails import BuiltInValidatorGuardrail
+from uipath.platform.guardrails.decorators import GuardrailExecutionStage
 from uipath.platform.guardrails.decorators._exceptions import GuardrailBlockException
 
 from uipath_langchain.agent.exceptions import AgentRuntimeError
 
 from ..models import GuardrailAction
-from ._utils import convert_block_exception, extract_text_from_messages
+from ._utils import (
+    convert_block_exception,
+    create_modified_tool_request,
+    create_modified_tool_result,
+    extract_text_from_messages,
+    sanitize_tool_name,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -35,6 +44,8 @@ class BuiltInGuardrailMiddlewareMixin:
     _guardrail: BuiltInValidatorGuardrail
     _name: str
     action: GuardrailAction
+    _tool_names: list[str] | None
+    _tool_stage: GuardrailExecutionStage
     _uipath: UiPath | None = None
 
     def _get_uipath(self) -> UiPath:
@@ -90,6 +101,91 @@ class BuiltInGuardrailMiddlewareMixin:
                 except (ValueError, SyntaxError):
                     return {"output": content}
         return {"output": content}
+
+    def _extract_tool_input_data(
+        self, request: ToolCallRequest
+    ) -> str | dict[str, Any]:
+        """Extract tool input data from ToolCallRequest for guardrail evaluation."""
+        tool_call = request.tool_call
+        args = tool_call.get("args", {})
+        if isinstance(args, dict):
+            return args
+        return str(args)
+
+    def _create_tool_wrap_hook(self, guardrail_name: str) -> AgentMiddleware:
+        """Create a wrap_tool_call hook that evaluates the guardrail at PRE and/or POST stage."""
+        middleware_instance = self
+
+        async def _wrap_tool_call_func(
+            request: ToolCallRequest,
+            handler: Any,
+        ) -> ToolMessage | Command[Any]:
+            tool_call = request.tool_call
+            tool_name = tool_call.get("name", "")
+            sanitized_tool_name = sanitize_tool_name(tool_name)
+
+            if (
+                middleware_instance._tool_names is None
+                or sanitized_tool_name not in middleware_instance._tool_names
+            ):
+                return await handler(request)
+
+            input_data = middleware_instance._extract_tool_input_data(request)
+
+            if middleware_instance._tool_stage in (
+                GuardrailExecutionStage.PRE,
+                GuardrailExecutionStage.PRE_AND_POST,
+            ):
+                try:
+                    result = middleware_instance._evaluate_guardrail(input_data)
+                    modified_input = middleware_instance._handle_validation_result(
+                        result, input_data
+                    )
+                    if modified_input is not None and isinstance(modified_input, dict):
+                        request = create_modified_tool_request(request, modified_input)
+                except GuardrailBlockException as exc:
+                    raise convert_block_exception(exc) from exc
+                except AgentRuntimeError:
+                    raise
+                except Exception as e:
+                    logger.error(
+                        f"Error evaluating '{middleware_instance._name}' guardrail (PRE)"
+                        f" for tool '{tool_name}': {e}",
+                        exc_info=True,
+                    )
+
+            tool_result = await handler(request)
+
+            if middleware_instance._tool_stage in (
+                GuardrailExecutionStage.POST,
+                GuardrailExecutionStage.PRE_AND_POST,
+            ):
+                output_data = middleware_instance._extract_tool_output_data(tool_result)
+                if output_data:
+                    try:
+                        result = middleware_instance._evaluate_guardrail(output_data)
+                        modified_output = middleware_instance._handle_validation_result(
+                            result, output_data
+                        )
+                        if modified_output is not None:
+                            tool_result = create_modified_tool_result(
+                                tool_result, modified_output
+                            )
+                    except GuardrailBlockException as exc:
+                        raise convert_block_exception(exc) from exc
+                    except AgentRuntimeError:
+                        raise
+                    except Exception as e:
+                        logger.error(
+                            f"Error evaluating '{middleware_instance._name}' guardrail (POST)"
+                            f" for tool '{tool_name}': {e}",
+                            exc_info=True,
+                        )
+
+            return tool_result
+
+        _wrap_tool_call_func.__name__ = f"{guardrail_name}_wrap_tool_call"
+        return wrap_tool_call(_wrap_tool_call_func)  # type: ignore[call-overload]
 
     def _check_messages(self, messages: list[BaseMessage]) -> None:
         """Evaluate guardrail against message text; apply action on violation."""
